requirements.txt
gunicorn
flask
flask-restful
flask-socketio==4.3.2
Werkzeug==2.0.3
flask-cors
celery
redis
psycopg2
gevent
gevent-websocket
wsaccel
python-ulid
